Improve C-only detection
We rely now on the detection of

del self.settings.compiler.libcxx
and
del self.settings.compiler.cppstd

in the recipe instead of analyzing the downloaded source code.

This is much, much, much faster and actually more reliable as source code can have C++ files in e.g. external dependencies or test files which are not part of the build.

This also reflects the current status of the Conan Center hooks.