Support testable header only and standalone recipes.
This adds support for testable header only recipes, i.e. that have a
test_package, to be tested in all the build configurations possible.

It also adds support for "standalone" recipes. Such recipes contain
what would normally be in the recipe build subdir, i.e. conanfile.py
and conandata.yml, but at the root level. This makes it easy for
library authors to create an in-source recipe with minimal effort.