Avoid detection collision between styles of recipes.
It's possible to have one-recipe-one-version with a conandata.yml. That
structure would conflict with the standalone-recipe-many-version. This
fixes that by using the existance of a recipe version from env, CI, or
the conanfile.py as a discriminator.

This also fixes an ambiguity on the fallback recipe version parsing from
conanfile.py when there's a required_conan_version, or other version
variable, specified in the conanfile.py.